README: Improve formatting, add warning for unsupported systems
This commit tries to improve markdown formatting of this project's README, while keeping it readable in plain text.

Some line breaks weren't showing up in markdown. This has been fixed. Code blocks were added to distinguish commands from the rest of the text. Some wording has also been changed slightly to make the documentation a bit easier to read.

Lastly, it clarifies that enablement must be done for each system (not just the WWAN module) and that unlisted systems are unsupported.