Ubuntu oem (#22)
* Enable WWAN for few ThinkPad Products
Below Modifications are done for this release:
1) Enabled WWAN for below products:
- (T16 Gen 3) Intel (CS24 Product)
- (T14 Gen 5) Intel/AMD (CS24 Product)
- (T14s Gen 5) Intel (CS24 Product)
- (X13 Gen 5) Intel (CS24 Product)
- (X13 Flex Gen 5) Intel (CS24 Product)
-  L14 Gen 4 (only for L860R+)
-  L15 Gen 4 (only for L860R+)

2) Fix Modem fail issue after reboot

3) Fix security issue

* Add script to uninstall fccunlock package

* fix issue releated to country checking

* feat(ModemManager.service): use drop-in override (#18)

* feat(ModemManager.service): use drop-in override

Instead of modifying /usr/lib/systemd/system/ModemManager.service
directly. Use a drop-in configuration to override the `ExecStart`. So,
even if users update the ModemManager with their package manager, the
override will not be affect, which means they don't need to rerun the
fcc_unlock_setup.sh.

A wrapper is introduced to check users' ModemManager's support scope at
runtime instead of installation time. So users can free to upgrade or
downgrade ModemManager without encounter the `unknown option` errors.

* fix(uninstall): remove the drop-in file

* Fix checking condition of module type

* Fix error

* Below changes are done:

- Added verbose log
- Move all libraries to /opt folder
- Code clean up

* Updated README

---------

Co-authored-by: Nitin Joshi <njoshi1@lenovo.com>
Co-authored-by: Nitin Joshi <nitjoshi@gmail.com>
Co-authored-by: Atlas Yu <85104110+pseudocc@users.noreply.github.com>